Core: refactor into subpackages
