+++ brevitas/core/bit_width/parameter.py
@@ -46,35 +46,11 @@ from torch.nn import Parameter, Module
 
 import brevitas
 import brevitas.config as config
-from brevitas.function.ops_ste import tensor_clamp_ste
-from brevitas.inject.enum import BitWidthImplType  # retrocompatibility
-from .utils import StatelessBuffer
-
-assert BitWidthImplType  # prevent removal of unused import
 
 
 MIN_INT_BIT_WIDTH = 2
 NON_ZERO_EPSILON = 1e-6
 REMOVE_ZERO_BIT_WIDTH = 0.1
-
-
-class IdentityBitWidth(brevitas.jit.ScriptModule):
-
-    @brevitas.jit.script_method
-    def forward(self, x: Tensor) -> Tensor:
-        return x
-
-
-class BitWidthConst(brevitas.jit.ScriptModule):
-
-    def __init__(self, bit_width: int) -> None:
-        super(BitWidthConst, self).__init__()
-        assert isinstance(bit_width, int)
-        self.bit_width = StatelessBuffer(torch.tensor(float(bit_width)))
-
-    @brevitas.jit.script_method
-    def forward(self) -> Tensor:
-        return self.bit_width()
 
 
 class BitWidthParameter(brevitas.jit.ScriptModule):
@@ -156,28 +132,6 @@ class RemoveBitwidthParameter(brevitas.jit.ScriptModule):
             missing_keys.remove(bit_width_coeff_key)
 
 
-class MsbClampBitWidth(brevitas.jit.ScriptModule):
 
-    def __init__(
-            self,
-            bit_width_to_remove_impl: Module,
-            min_overall_bit_width: int,
-            max_overall_bit_width: int) -> None:
-        super(MsbClampBitWidth, self).__init__()
-
-        self.min_overall_bit_width = BitWidthConst(min_overall_bit_width)
-        self.max_overall_bit_width = BitWidthConst(max_overall_bit_width)
-        self.bit_width_to_remove_impl = bit_width_to_remove_impl
-
-    @brevitas.jit.script_method
-    def forward(self, input_bit_width: Tensor) -> Tensor:
-        bit_width_to_remove = self.bit_width_to_remove_impl()
-        output_bit_width = torch.abs(input_bit_width - bit_width_to_remove)
-        # todo STE on max only
-        output_bit_width = tensor_clamp_ste(
-            output_bit_width,
-            self.min_overall_bit_width(),
-            self.max_overall_bit_width())
-        return output_bit_width
