+++ brevitas/core/function_wrapper/shape.py
@@ -40,136 +40,8 @@
 
 import brevitas
 import torch
-from brevitas.function.ops_ste import round_ste, tensor_clamp_ste, ceil_ste, floor_ste, round_to_zero_ste
-from brevitas.function.shape import *
-from brevitas.function import tensor_clamp, identity
-
-
-class Identity(brevitas.jit.ScriptModule):
-    def __init__(self) -> None:
-        super(Identity, self).__init__()
-
-    @brevitas.jit.script_method
-    def forward(self, x: torch.Tensor):
-        return identity(x)
-
-
-class RoundSte(brevitas.jit.ScriptModule):
-    def __init__(self) -> None:
-        super(RoundSte, self).__init__()
-
-    @brevitas.jit.script_method
-    def forward(self, x: torch.Tensor):
-        return round_ste(x)
-
-
-class FloorSte(brevitas.jit.ScriptModule):
-    def __init__(self) -> None:
-        super(FloorSte, self).__init__()
-
-    @brevitas.jit.script_method
-    def forward(self, x: torch.Tensor):
-        return floor_ste(x)
-
-
-class RoundToZeroSte(brevitas.jit.ScriptModule):
-    def __init__(self) -> None:
-        super(RoundToZeroSte, self).__init__()
-
-    @brevitas.jit.script_method
-    def forward(self, x: torch.Tensor):
-        return round_to_zero_ste(x)
-
-
-class CeilSte(brevitas.jit.ScriptModule):
-    def __init__(self) -> None:
-        super(CeilSte, self).__init__()
-
-    @brevitas.jit.script_method
-    def forward(self, x: torch.Tensor):
-        return ceil_ste(x)
-
-
-class PowerOfTwo(brevitas.jit.ScriptModule):
-    def __init__(self) -> None:
-        super(PowerOfTwo, self).__init__()
-
-    @brevitas.jit.script_method
-    def forward(self, x: torch.Tensor):
-        return 2.0 ** x
-
-
-class LogTwo(brevitas.jit.ScriptModule):
-    def __init__(self) -> None:
-        super(LogTwo, self).__init__()
-
-    @brevitas.jit.script_method
-    def forward(self, x: torch.Tensor):
-        return torch.log2(x)
-
-
-class InplaceLogTwo(brevitas.jit.ScriptModule):
-    def __init__(self) -> None:
-        super(InplaceLogTwo, self).__init__()
-
-    @brevitas.jit.script_method
-    def forward(self, x: torch.Tensor):
-        return x.log2_()
-
-
-class TensorClampSte(brevitas.jit.ScriptModule):
-    def __init__(self) -> None:
-        super(TensorClampSte, self).__init__()
-
-    @brevitas.jit.script_method
-    def forward(self, x: torch.Tensor, min_val: torch.Tensor, max_val: torch.Tensor):
-        return tensor_clamp_ste(x, min_val, max_val)
-
-
-class TensorClamp(brevitas.jit.ScriptModule):
-    def __init__(self) -> None:
-        super(TensorClamp, self).__init__()
-
-    @brevitas.jit.script_method
-    def forward(self, x: torch.Tensor, min_val: torch.Tensor, max_val: torch.Tensor):
-        return tensor_clamp(x, min_val=min_val, max_val=max_val)
-
-
-class ConstScalarClamp(brevitas.jit.ScriptModule):
-    __constants__ = ['min_val', 'max_val']
-
-    def __init__(self, min_val, max_val) -> None:
-        super(ConstScalarClamp, self).__init__()
-        self.min_val = min_val
-        self.max_val = max_val
-
-    @brevitas.jit.script_method
-    def forward(self, x: torch.Tensor):
-        return torch.clamp(x, min=self.min_val, max=self.max_val)
-
-
-class ClampMin(brevitas.jit.ScriptModule):
-    __constants__ = ['min_val']
-
-    def __init__(self, min_val: float) -> None:
-        super(ClampMin, self).__init__()
-        self.min_val = min_val
-
-    @brevitas.jit.script_method
-    def forward(self, x: torch.Tensor):
-        return x.clamp_min(self.min_val)
-
-
-class ClampMinSte(brevitas.jit.ScriptModule):
-    __constants__ = ['min_val']
-
-    def __init__(self, min_val: float) -> None:
-        super(ClampMinSte, self).__init__()
-        self.min_val = min_val
-
-    @brevitas.jit.script_method
-    def forward(self, x: torch.Tensor):
-        return x.clamp_min(self.min_val)
+from brevitas.function.shape import over_tensor, over_output_channels, over_batch_over_tensor
+from brevitas.function.shape import over_batch_over_output_channels
 
 
 class OverTensorView(brevitas.jit.ScriptModule):
