+++ brevitas/core/quant/delay.py
@@ -1,0 +1,90 @@
+# Copyright (c) 2018-     Xilinx, Inc              (Alessandro Pappalardo)
+# Copyright (c) 2016-     Facebook, Inc            (Adam Paszke)
+# Copyright (c) 2014-     Facebook, Inc            (Soumith Chintala)
+# Copyright (c) 2011-2014 Idiap Research Institute (Ronan Collobert)
+# Copyright (c) 2012-2014 Deepmind Technologies    (Koray Kavukcuoglu)
+# Copyright (c) 2011-2012 NEC Laboratories America (Koray Kavukcuoglu)
+# Copyright (c) 2011-2013 NYU                      (Clement Farabet)
+# Copyright (c) 2006-2010 NEC Laboratories America (Ronan Collobert, Leon Bottou, Iain Melvin, Jason Weston)
+# Copyright (c) 2006      Idiap Research Institute (Samy Bengio)
+# Copyright (c) 2001-2004 Idiap Research Institute (Ronan Collobert, Samy Bengio, Johnny Mariethoz)
+
+# All rights reserved.
+
+# Redistribution and use in source and binary forms, with or without
+# modification, are permitted provided that the following conditions are met:
+
+# 1. Redistributions of source code must retain the above copyright
+#    notice, this list of conditions and the following disclaimer.
+
+# 2. Redistributions in binary form must reproduce the above copyright
+#    notice, this list of conditions and the following disclaimer in the
+#    documentation and/or other materials provided with the distribution.
+
+# 3. Neither the names of Xilinx, Facebook, Deepmind Technologies, NYU,
+#    NEC Laboratories America and IDIAP Research Institute nor the names
+#    of its contributors may be used to endorse or promote products derived
+#    from this software without specific prior written permission.
+
+# THIS SOFTWARE IS PROVIDED BY THE COPYRIGHT HOLDERS AND CONTRIBUTORS "AS IS"
+# AND ANY EXPRESS OR IMPLIED WARRANTIES, INCLUDING, BUT NOT LIMITED TO, THE
+# IMPLIED WARRANTIES OF MERCHANTABILITY AND FITNESS FOR A PARTICULAR PURPOSE
+# ARE DISCLAIMED. IN NO EVENT SHALL THE COPYRIGHT OWNER OR CONTRIBUTORS BE
+# LIABLE FOR ANY DIRECT, INDIRECT, INCIDENTAL, SPECIAL, EXEMPLARY, OR
+# CONSEQUENTIAL DAMAGES (INCLUDING, BUT NOT LIMITED TO, PROCUREMENT OF
+# SUBSTITUTE GOODS OR SERVICES; LOSS OF USE, DATA, OR PROFITS; OR BUSINESS
+# INTERRUPTION) HOWEVER CAUSED AND ON ANY THEORY OF LIABILITY, WHETHER IN
+# CONTRACT, STRICT LIABILITY, OR TORT (INCLUDING NEGLIGENCE OR OTHERWISE)
+# ARISING IN ANY WAY OUT OF THE USE OF THIS SOFTWARE, EVEN IF ADVISED OF THE
+# POSSIBILITY OF SUCH DAMAGE.
+
+from typing import Optional
+
+from torch import Tensor
+
+import brevitas
+
+
+class _NoDelay(brevitas.jit.ScriptModule):
+
+    @brevitas.jit.script_method
+    def forward(self, x: Tensor, y: Tensor) -> Tensor:
+        return y
+
+
+class _DelayQuant(brevitas.jit.ScriptModule):
+
+    def __init__(self, quant_delay_steps):
+        super(_DelayQuant, self).__init__()
+        self.quant_delay_steps: int = brevitas.jit.Attribute(quant_delay_steps, int)
+
+    @brevitas.jit.script_method_110_disabled
+    def forward(self, x: Tensor, y: Tensor) -> Tensor:
+        if self.training and self.quant_delay_steps > 0:
+            self.quant_delay_steps = self.quant_delay_steps - 1
+            return x
+        else:
+            return y
+
+    def _load_from_state_dict(self, state_dict, prefix, local_metadata, strict,
+                              missing_keys, unexpected_keys, error_msgs):
+        super(_DelayQuant, self)._load_from_state_dict(
+            state_dict, prefix, local_metadata, strict, missing_keys, unexpected_keys, error_msgs)
+        # Pytorch stores training flag as a buffer with JIT enabled
+        training_key = prefix + 'training'
+        if training_key in missing_keys:
+            missing_keys.remove(training_key)
+
+
+class DelayWrapper(brevitas.jit.ScriptModule):
+
+    def __init__(self, quant_delay_steps: Optional[int]):
+        super(DelayWrapper, self).__init__()
+        if quant_delay_steps is None or quant_delay_steps <= 0:
+            self.delay_impl = _NoDelay()
+        else:
+            self.delay_impl = _DelayQuant(quant_delay_steps)
+
+    @brevitas.jit.script_method
+    def forward(self, x: Tensor, y: Tensor) -> Tensor:
+        return self.delay_impl(x, y)
